+++ bittytax/conv/parsers/cryptocom.py
@@ -147,6 +147,14 @@ DataParser(DataParser.TYPE_EXCHANGE,
            "Crypto.com",
            ['Timestamp (UTC)', 'Transaction Description', 'Currency', 'Amount', 'To Currency',
             'To Amount', 'Native Currency', 'Native Amount', 'Native Amount (in USD)',
+            'Transaction Kind', 'Transaction Hash'],
+           worksheet_name="Crypto.com",
+           row_handler=parse_crypto_com)
+
+DataParser(DataParser.TYPE_EXCHANGE,
+           "Crypto.com",
+           ['Timestamp (UTC)', 'Transaction Description', 'Currency', 'Amount', 'To Currency',
+            'To Amount', 'Native Currency', 'Native Amount', 'Native Amount (in USD)',
             'Transaction Kind'],
            worksheet_name="Crypto.com",
            row_handler=parse_crypto_com)