+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - Conversion tool: added parser for CoinList exchange.
 - Conversion tool: added parser for Koinly accounting data.
 - Accounting tool: tax rates and allowance for 2022/23.
+- Crypto.com parser: added new export format. ([#248](https://github.com/BittyTax/BittyTax/issues/248))
 ### Changed
 - Binance parser: performance improvements for large data sets.
 - Coinbase Pro parser: performance improvements for large data sets.